+++ src/main/java/kitchenpos/order/validator/OrderValidator.java
@@ -1,40 +1,15 @@
 package kitchenpos.order.validator;
 
 import java.util.List;
-import kitchenpos.common.constant.ErrorCode;
-import kitchenpos.common.exception.NotFoundException;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.dto.OrderRequest;
 import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.repository.OrderTableRepository;
-import org.springframework.stereotype.Component;
 
-@Component
-public class OrderValidator {
+public interface OrderValidator {
 
-    private final OrderTableRepository orderTableRepository;
+    void validateNotCompleteOrders(List<Order> orders);
 
-    public OrderValidator(OrderTableRepository orderTableRepository) {
-        this.orderTableRepository = orderTableRepository;
-    }
+    void validatorOrder(OrderRequest orderRequest);
 
-    public void validator(OrderRequest orderRequest) {
-        OrderTable orderTable = findOrderTableById(orderRequest.getOrderTableId());
-        validateOrderTable(orderTable);
-    }
-
-    private OrderTable findOrderTableById(Long id) {
-        return orderTableRepository.findById(id)
-            .orElseThrow(() -> new NotFoundException());
-    }
-
-    private void validateOrderTable(OrderTable orderTable) {
-        if (orderTable.isEmpty()) {
-            throw new IllegalArgumentException(ErrorCode.ORDER_TABLE_NOT_EMPTY.getErrorMessage());
-        }
-    }
-
-    public void validateNotCompleteOrders(List<Order> orders) {
-        orders.forEach(Order::validateNotCompleteOrder);
-    }
+    void validatorTable(OrderTable orderTable);
 }
